requirements.txt
coreapi==2.1.1
dj-database-url==0.4.1
Django==1.10.4
djangorestframework==3.5.3
drfdocs==0.0.11
gunicorn==19.6.0
itypes==1.1.0
openapi-codec==1.2.1
Pillow==3.4.2
psycopg2==2.6.2
requests==2.12.3
simplejson==3.10.0
uritemplate==3.0.0
whitenoise==3.2.2
django-cors-headers==2.0.2
social-auth-app-django==1.1.0
social-auth-core==1.2.0

fcm-django==0.2.6